Unite documentCache and entityCache into apiCache
Assuming their id spaces will never overlap.
