+++ ui/src/components/DocumentScreen/index.jsx
@@ -36,7 +36,7 @@ class DocumentScreen extends Component {
 const mapStateToProps = (state, ownProps) => {
   const { documentId } = ownProps.match.params;
   const document = documentId !== undefined
-    ? state.documentCache[documentId]
+    ? state.apiCache[documentId]
     : undefined;
   return { documentId, document };
 }
